requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.6.2
numpy==1.24.0
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-pptx==0.6.21
pytz==2022.7
redis==4.4.0
requests==2.28.1
seaborn==0.12.1
six==1.16.0
spotipy==2.22.0
urllib3==1.26.13
XlsxWriter==3.0.3
